+++ docs/source/changelog.rst
@@ -25,6 +25,9 @@ Added
 * Add |Flask| as an extra dependency
 * Add |CurrentMessage| Middleware that exposes the current message via a thread local variable, useful to access the message within the actor
 * Add |CurrentMessage| to the list of ``default_middleware``
+* Add new POST method ``cancel_message`` in module ``api``
+   - url ``/messages/cancel/message_id``
+
 `0.20.0`_ -- 2020-04-07
 -----------------------
 BREAKING CHANGE